improve: reorganize playback structure and worker files
- Moved playback structures to src/playback/structs/
- Moved playback processing logic to src/playback/processing/
- Moved worker entry points to src/workers/
- Updated imports and paths across the codebase to reflect new structure
