+++ src/playback/processing/filtersManager.js
@@ -1,23 +1,23 @@
 import { Transform } from 'node:stream'
-import { logger } from '../utils.js'
+import { logger } from '../../utils.js'
 
-import ChannelMix from './filters/channelMix.js'
-import Chorus from './filters/chorus.js'
-import Compressor from './filters/compressor.js'
-import Distortion from './filters/distortion.js'
-import Echo from './filters/echo.js'
-import Equalizer from './filters/equalizer.js'
-import Flanger from './filters/flanger.js'
-import Highpass from './filters/highpass.js'
-import Karaoke from './filters/karaoke.js'
-import Lowpass from './filters/lowpass.js'
-import Phaser from './filters/phaser.js'
-import Reverb from './filters/reverb.js'
-import Rotation from './filters/rotation.js'
-import Spatial from './filters/spatial.js'
-import Timescale from './filters/timescale.js'
-import Tremolo from './filters/tremolo.js'
-import Vibrato from './filters/vibrato.js'
+import ChannelMix from '../filters/channelMix.js'
+import Chorus from '../filters/chorus.js'
+import Compressor from '../filters/compressor.js'
+import Distortion from '../filters/distortion.js'
+import Echo from '../filters/echo.js'
+import Equalizer from '../filters/equalizer.js'
+import Flanger from '../filters/flanger.js'
+import Highpass from '../filters/highpass.js'
+import Karaoke from '../filters/karaoke.js'
+import Lowpass from '../filters/lowpass.js'
+import Phaser from '../filters/phaser.js'
+import Reverb from '../filters/reverb.js'
+import Rotation from '../filters/rotation.js'
+import Spatial from '../filters/spatial.js'
+import Timescale from '../filters/timescale.js'
+import Tremolo from '../filters/tremolo.js'
+import Vibrato from '../filters/vibrato.js'
 
 const FILTER_CLASSES = {
   tremolo: Tremolo,
