+++ src/playback/processing/streamProcessor.js
@@ -6,13 +6,13 @@ import FAAD2NodeDecoder from '@ecliptia/faad2-wasm/faad2_node_decoder.js'
 import { SeekError, seekableStream } from '@ecliptia/seekable-stream'
 import { SymphoniaDecoder } from '@toddynnn/symphonia-decoder'
 import * as MP4Box from 'mp4box'
-import { normalizeFormat, SupportedFormats } from '../constants.js'
-import { logger } from '../utils.js'
-import FlvDemuxer from './demuxers/Flv.js'
-import WebmOpusDemuxer from './demuxers/WebmOpus.js'
+import { normalizeFormat, SupportedFormats } from '../../constants.js'
+import { logger } from '../../utils.js'
+import FlvDemuxer from '../demuxers/Flv.js'
+import WebmOpusDemuxer from '../demuxers/WebmOpus.js'
 import { FiltersManager } from './filtersManager.js'
-import { Decoder as OpusDecoder, Encoder as OpusEncoder } from './opus/Opus.js'
-import { RingBuffer } from './RingBuffer.js'
+import { Decoder as OpusDecoder, Encoder as OpusEncoder } from '../opus/Opus.js'
+import { RingBuffer } from '../structs/RingBuffer.js'
 import { VolumeTransformer } from './VolumeTransformer.js'
 
 const AUDIO_CONFIG = Object.freeze({
