+++ src/playback/structs/BufferPool.js
@@ -1,4 +1,4 @@
-import { logger } from '../utils.js'
+import { logger } from '../../utils.js'
 
 const MAX_POOL_SIZE_BYTES = 50 * 1024 * 1024
 const CLEANUP_INTERVAL = 60000
