+++ src/workers/main.js
@@ -2,20 +2,20 @@ import net from 'node:net'
 import os from 'node:os'
 import { monitorEventLoopDelay } from 'node:perf_hooks'
 import v8 from 'node:v8'
-import { GatewayEvents } from './constants.js'
-import ConnectionManager from './managers/connectionManager.js'
-import CredentialManager from './managers/credentialManager.js'
-import TrackCacheManager from './managers/trackCacheManager.js'
-import LyricsManager from './managers/lyricsManager.js'
-import PluginManager from './managers/pluginManager.js'
-import RoutePlannerManager from './managers/routePlannerManager.js'
-import SourceManager from './managers/sourceManager.js'
-import StatsManager from './managers/statsManager.js'
-import { bufferPool } from './playback/BufferPool.js'
-import { Player } from './playback/player.js'
-import { createPCMStream } from './playback/streamProcessor.js'
-import { cleanupHttpAgents, initLogger, logger } from './utils.js'
-import { createVoiceRelay } from './voice/voiceRelay.js'
+import { GatewayEvents } from '../constants.js'
+import ConnectionManager from '../managers/connectionManager.js'
+import CredentialManager from '../managers/credentialManager.js'
+import TrackCacheManager from '../managers/trackCacheManager.js'
+import LyricsManager from '../managers/lyricsManager.js'
+import PluginManager from '../managers/pluginManager.js'
+import RoutePlannerManager from '../managers/routePlannerManager.js'
+import SourceManager from '../managers/sourceManager.js'
+import StatsManager from '../managers/statsManager.js'
+import { bufferPool } from '../playback/structs/BufferPool.js'
+import { Player } from '../playback/player.js'
+import { createPCMStream } from '../playback/processing/streamProcessor.js'
+import { cleanupHttpAgents, initLogger, logger } from '../utils.js'
+import { createVoiceRelay } from '../voice/voiceRelay.js'
 
 let lastCpuUsage = process.cpuUsage()
 let lastCpuTime = Date.now()
@@ -35,9 +35,9 @@ try {
 
 let config
 try {
-  config = (await import('../config.js')).default
+  config = (await import('../../config.js')).default
 } catch {
-  config = (await import('../config.default.js')).default
+  config = (await import('../../config.default.js')).default
 }
 
 const HIBERNATION_ENABLED = config.cluster?.hibernation?.enabled !== false
@@ -258,7 +258,7 @@ nodelink.pluginManager = new PluginManager(nodelink)
 nodelink.registry = null
 if (process.embedder === 'nodejs') {
   try {
-    nodelink.registry = await import('./registry.js')
+    nodelink.registry = await import('../registry.js')
   } catch (e) {
     logger('error', 'Worker', `Failed to load registry: ${e.message}`)
   }
